// ui/src/api/gen/StatePayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppSettings } from "./AppSettings";
import type { Profile } from "./Profile";
import type { ProfileMeta } from "./ProfileMeta";

export type StatePayload = { settings: AppSettings, profile: Profile, profiles: Array<ProfileMeta>, };